Remove old feature gates from pre-4.0
