+++ cmd/src/batch_remote.go
@@ -52,7 +52,7 @@ Examples:
 			Client: cfg.apiClient(flags.api, flagSet.Output()),
 		})
 
-		if err := svc.DetermineFeatureFlags(ctx); err != nil {
+		if err := validateSourcegraphVersionConstraint(ctx, svc); err != nil {
 			return err
 		}
