+++ cmd/src/batch_validate.go
@@ -63,7 +63,7 @@ Examples:
 			Client: cfg.apiClient(apiFlags, flagSet.Output()),
 		})
 
-		if err := svc.DetermineFeatureFlags(ctx); err != nil {
+		if err := validateSourcegraphVersionConstraint(ctx, svc); err != nil {
 			ui.ExecutionError(err)
 			return err
 		}
