--- internal/batches/graphql/batch_change.go
@@ -1,5 +1,0 @@
-package graphql
-
-type BatchChange struct {
-	URL string
-}
