+++ internal/batches/service/remote.go
@@ -8,13 +8,6 @@ import (
 
 var ErrServerSideBatchChangesUnsupported = errors.New("server side batch changes are not available on this Sourcegraph instance")
 
-func (svc *Service) areServerSideBatchChangesSupported() error {
-	if !svc.features.ServerSideBatchChanges {
-		return ErrServerSideBatchChangesUnsupported
-	}
-	return nil
-}
-
 const upsertEmptyBatchChangeQuery = `
 mutation UpsertEmptyBatchChange(
 	$name: String!
@@ -34,10 +27,6 @@ func (svc *Service) UpsertBatchChange(
 	name string,
 	namespaceID string,
 ) (string, error) {
-	if err := svc.areServerSideBatchChangesSupported(); err != nil {
-		return "", err
-	}
-
 	var resp struct {
 		UpsertEmptyBatchChange struct {
 			Name string `json:"name"`
@@ -82,10 +71,6 @@ func (svc *Service) CreateBatchSpecFromRaw(
 	allowUnsupported bool,
 	noCache bool,
 ) (string, error) {
-	if err := svc.areServerSideBatchChangesSupported(); err != nil {
-		return "", err
-	}
-
 	var resp struct {
 		CreateBatchSpecFromRaw struct {
 			ID string `json:"id"`
@@ -118,10 +103,6 @@ func (svc *Service) ExecuteBatchSpec(
 	batchSpecID string,
 	noCache bool,
 ) (string, error) {
-	if err := svc.areServerSideBatchChangesSupported(); err != nil {
-		return "", err
-	}
-
 	var resp struct {
 		ExecuteBatchSpec struct {
 			ID string `json:"id"`
@@ -157,10 +138,6 @@ type BatchSpecWorkspaceResolution struct {
 }
 
 func (svc *Service) GetBatchSpecWorkspaceResolution(ctx context.Context, id string) (*BatchSpecWorkspaceResolution, error) {
-	if err := svc.areServerSideBatchChangesSupported(); err != nil {
-		return nil, err
-	}
-
 	var resp struct {
 		Node struct {
 			WorkspaceResolution BatchSpecWorkspaceResolution `json:"workspaceResolution"`
